requirements.txt
Flask
Flask-SocketIO
Flask-Cors
Flask-SQLAlchemy
python-dotenv